Extracted util
